Fix freebsd/386
